refactoring before factory-function
